@@ -3,8 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
+    "dotenv": "^8.2.0",
+    "react": "^16.13.0",
+    "react-dom": "^16.13.0",
     "react-scripts": "1.0.14"
   },
   "scripts": {
