@@ -1,5 +1,8 @@
+import dotenv from 'dotenv';
+dotenv.config();
+
 const TMDB = {
-  api_key: '<REPLACE_THIS_WITH_TMDB_API_KEY>',
+  api_key: process.env.API_KEY,
   films: [
     {
       "id": 346364,
